// api/model/covid19Info.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Covid19Info { 
    diagnose?: string;
    rkiCrit?: Covid19Info.RkiCritEnum;
    rkiReason?: string;
}
export namespace Covid19Info {
    export type RkiCritEnum = 'NOTHING' | 'RKI1' | 'RKI2' | 'RKI3' | 'RKI4';
    export const RkiCritEnum = {
        NOTHING: 'NOTHING' as RkiCritEnum,
        RKI1: 'RKI1' as RkiCritEnum,
        RKI2: 'RKI2' as RkiCritEnum,
        RKI3: 'RKI3' as RkiCritEnum,
        RKI4: 'RKI4' as RkiCritEnum
    };
}