// api/model/formDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Covid19Info } from './covid19Info';
import { OrdererDto } from './ordererDto';
import { PatientDto } from './patientDto';

export interface FormDto { 
    orderer?: OrdererDto;
    patient?: PatientDto;
    info?: Covid19Info;
}